backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS support
fastapi-cors==0.0.6

# Data models and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Embeddings and ML
sentence-transformers==2.2.2
torch>=2.0.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
mammoth==1.6.0

# Database support
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Vector database
qdrant-client==1.7.0

# Caching
redis==5.0.1

# Async support
aiofiles==23.2.1
asyncio==3.4.3

# Logging and monitoring
python-json-logger==2.0.7

# Optional: For enhanced NLP
spacy==3.7.2
# Run: python -m spacy download en_core_web_sm

# Optional: For LLM integration
# openai==1.3.0
# langchain==0.0.340
# anthropic==0.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

google-generativeai==0.8.3
protobuf<5